refactor: usePageResize 커스텀 훅 생성
